Updated data type of this class

+++ src/edu/moumita/practise/CheckNumberIsEvenOROdd.java
@@ -6,7 +6,7 @@ public class CheckNumberIsEvenOROdd {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         System.out.println("Enter a number : ");
-        int num = scanner.nextInt();
+        float num = scanner.nextFloat();
 
         if(num%2==0){
             System.out.println("The number is even");
